+++ src/math/extgcd.hpp
@@ -1,6 +1,6 @@
 #include "../template/template.hpp"
-ll extgcd(ll a, ll b, ll& x, ll& y) {
-    ll d = a;
+long long extgcd(long long a, long long b, long long& x, long long& y) {
+    long long d = a;
     if(b != 0) {
         d = extgcd(b, a % b, y, x);
         y -= (a / b) * x;
