+++ src/math/inv_gcd.hpp
@@ -1,15 +1,15 @@
 #pragma once
 #include "../template/template.hpp"
-constexpr pair<ll, ll> inv_gcd(ll a, ll b) {
+constexpr pair<long long, long long> inv_gcd(long long a, long long b) {
     a %= b;
     if(a < 0) a += b;
     if(a == 0) return {b, 0};
-    ll s = b, t = a, m0 = 0, m1 = 1;
+    long long s = b, t = a, m0 = 0, m1 = 1;
     while(t) {
-        const ll u = s / t;
+        const long long u = s / t;
         s -= t * u;
         m0 -= m1 * u;
-        ll tmp = s;
+        long long tmp = s;
         s = t;
         t = tmp;
         tmp = m0;
