+++ src/math/is_prime.hpp
@@ -1,8 +1,8 @@
 #pragma once
 #include "../template/template.hpp"
-constexpr bool is_prime(ll n) {
+constexpr bool is_prime(long long n) {
     if(n <= 1) return false;
-    for(ll i = 2; i * i <= n; ++i) {
+    for(long long i = 2; i * i <= n; ++i) {
         if(n % i == 0) return false;
     }
     return true;
